[addons][imagedecoder] separate "C" and "C++" parts in different files
There becomes image_decoder.h added where then include all "C" parts for
works by Kodi. The C++ part stays in ImageDecoder.h where used by addon.

Currently is the C++ header still included in Kodi, to have compile checks
during changes about, about them comes in other requests a test system
and Kodi only need to have the "C" parts of system.
